Add files via upload
This is the website created by my team for the 2022 BE Smart Hackathon. It is a website with features to ease anxiety for travelers by using interactive checklists and maps to ensure the traveler has everything they will need for their trip and does not get lost. My 5-person team had 24 hours to build this site with zero experience in HTML, CSS, or JavaScript.